backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
emails==0.6
jinja2==3.1.3
aiofiles==23.2.1
slowapi==0.1.9
websockets==12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
elasticsearch[async]==8.11.1
redis==5.0.1
motor==3.3.2

# Message Queue
confluent-kafka==2.3.0
aiokafka==0.10.0

# Security Tools
pymetasploit3==1.0.3
wazuhi==4.7.0
nuclei-sdk==0.1.0
yara-python==4.3.1
suricatasc==2.0.1

# Machine Learning
torch==2.1.2
transformers==4.36.2
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
sentry-sdk==1.39.1

# Logging
python-json-logger==2.0.7
structlog==23.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
hypothesis==6.92.1
faker==22.5.1

# Development
black==23.12.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocstrings==0.24.0 